requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
drf-nested-routers>=0.94.0
django-cors-headers>=4.3.0

# Authentication
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.9

# WebSocket Support
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# QR Code Generation
qrcode>=7.4.2
Pillow>=10.1.0

# Reporting
reportlab>=4.0.0
openpyxl>=3.1.2
xlsxwriter>=3.1.9

# Utilities
python-decouple>=3.8
django-filter>=23.5
drf-spectacular>=0.27.2
requests>=2.31.0
python-dateutil>=2.8.2

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0

# Development
django-debug-toolbar>=4.2.0
pytest>=7.4.0
pytest-django>=4.7.0
